requirements.txt
pyTelegramBotAPI
python-dotenv
google-generativeai
openai
flask
requests
ddgs
pypdf
python-docx
openpyxl
phpserialize
xlrd
